fetch_data/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
idna==3.4
psycopg2-binary==2.9.6
PyJWT==1.7.1
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==1.26.7
